# src/Web.Ui/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "WageCore",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/salary-decrees": {
      "post": {
        "tags": [
          "SalaryDecree"
        ],
        "summary": "CreateSalaryDecree",
        "operationId": "CreateSalaryDecree",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateSalaryDecreeRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateSalaryDecreeRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateSalaryDecreeRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSalaryDecreeCommandResponseResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSalaryDecreeCommandResponseResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSalaryDecreeCommandResponseResult"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "SalaryDecree"
        ],
        "summary": "GetSalaryDecrees",
        "operationId": "GetSalaryDecrees",
        "parameters": [
          {
            "name": "Pagination.PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Pagination.PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Pagination.Offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "EmployeeId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Status",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "x-enumNames": [
                "Active",
                "Expired"
              ],
              "enum": [
                0,
                1
              ]
            }
          },
          {
            "name": "WorkshopId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "DepartmentId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetSalaryDecreesResponsePagedResultResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSalaryDecreesResponsePagedResultResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSalaryDecreesResponsePagedResultResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/salary-decrees/{salaryDecreeId}": {
      "put": {
        "tags": [
          "SalaryDecree"
        ],
        "summary": "UpdateSalaryDecree",
        "operationId": "UpdateSalaryDecree",
        "parameters": [
          {
            "name": "salaryDecreeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UpdateSalaryDecreeRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UpdateSalaryDecreeRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UpdateSalaryDecreeRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "SalaryDecree"
        ],
        "summary": "DeleteSalaryDecree",
        "operationId": "DeleteSalaryDecree",
        "parameters": [
          {
            "name": "employeeId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "salaryDecreeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/salary-decrees/{salaryDecreeId}/edit": {
      "get": {
        "tags": [
          "SalaryDecree"
        ],
        "summary": "GetSalaryDecreeForEdit",
        "operationId": "GetSalaryDecreeForEdit",
        "parameters": [
          {
            "name": "salaryDecreeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetSalaryDecreeForEditResponseResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSalaryDecreeForEditResponseResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSalaryDecreeForEditResponseResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BooleanResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "badResultType": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "x-enumNames": [
              "NotFound",
              "Validation",
              "General",
              "Unauthorized"
            ],
            "enum": [
              0,
              1,
              2,
              3
            ]
          },
          "response": {
            "type": "boolean"
          },
          "errors": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "CreateSalaryDecreeCommandResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "salaryDecreeId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "CreateSalaryDecreeCommandResponseResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "badResultType": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "x-enumNames": [
              "NotFound",
              "Validation",
              "General",
              "Unauthorized"
            ],
            "enum": [
              0,
              1,
              2,
              3
            ]
          },
          "response": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/CreateSalaryDecreeCommandResponse"
              }
            ]
          },
          "errors": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "CreateSalaryDecreeRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "employeeId": {
            "type": "string",
            "format": "uuid"
          },
          "salaryProfile": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/SalaryDecreeRequest"
              }
            ]
          }
        }
      },
...
      "SalaryDecreeRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "effectiveFrom": {
            "type": "string"
          },
          "baseDailySalary": {
            "type": "number",
            "format": "double"
          },
          "attractionAllowance": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "supervisionAllowance": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "shiftType": {
            "type": "integer",
            "format": "int32",
            "x-enumNames": [
              "None",
              "MorningEvening",
              "MorningNight",
              "EveningNight",
              "MorningEveningNight"
            ],
            "enum": [
              0,
              1,
              2,
              3,
              4
            ]
          },
          "contractType": {
            "type": "integer",
            "format": "int32",
            "x-enumNames": [
              "Permanent",
              "Temporary",
              "FixedTerm"
            ],
            "enum": [
              1,
              2,
              3
            ]
          },
          "housingAllowance": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "foodAllowance": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "transportationAllowanceNet": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "karanehAmountNet": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        }
      },
...
      "GetSalaryDecreeForEditResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "salaryDecreeId": {
            "type": "string",
            "format": "uuid"
          },
          "employeeId": {
            "type": "string",
            "format": "uuid"
          },
          "effectiveFrom": {
            "type": "string",
            "nullable": true
          },
          "baseDailySalary": {
            "type": "number",
            "format": "double"
          },
          "attractionAllowance": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "supervisionAllowance": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "shiftType": {
            "type": "integer",
            "format": "int32",
            "x-enumNames": [
              "None",
              "MorningEvening",
              "MorningNight",
              "EveningNight",
              "MorningEveningNight"
            ],
            "enum": [
              0,
              1,
              2,
              3,
              4
            ]
          },
          "contractType": {
            "type": "integer",
            "format": "int32",
            "x-enumNames": [
              "Permanent",
              "Temporary",
              "FixedTerm"
            ],
            "enum": [
              1,
              2,
              3
            ]
          },
          "housingAllowance": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "foodAllowance": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "transportationAllowanceNet": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "karanehAmountNet": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        }
      },
      "GetSalaryDecreeForEditResponseResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "badResultType": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "x-enumNames": [
              "NotFound",
              "Validation",
              "General",
              "Unauthorized"
            ],
            "enum": [
              0,
              1,
              2,
              3
            ]
          },
          "response": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/GetSalaryDecreeForEditResponse"
              }
            ]
          },
          "errors": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "GetSalaryDecreesResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "salaryDecreeId": {
            "type": "string",
            "format": "uuid"
          },
          "employeeId": {
            "type": "string",
            "format": "uuid"
          },
          "employeeName": {
            "type": "string",
            "nullable": true
          },
          "personalCode": {
            "type": "string",
            "nullable": true
          },
          "workshopName": {
            "type": "string",
            "nullable": true
          },
          "departmentName": {
            "type": "string",
            "nullable": true
          },
          "displayEffectiveFrom": {
            "type": "string",
            "nullable": true
          },
          "baseDailySalary": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "x-enumNames": [
              "Active",
              "Expired"
            ],
            "enum": [
              0,
              1
            ]
          }
        }
      },
      "GetSalaryDecreesResponsePagedResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/GetSalaryDecreesResponse"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "readOnly": true,
            "format": "int32"
          }
        }
      },
      "GetSalaryDecreesResponsePagedResultResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "badResultType": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "x-enumNames": [
              "NotFound",
              "Validation",
              "General",
              "Unauthorized"
            ],
            "enum": [
              0,
              1,
              2,
              3
            ]
          },
          "response": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/GetSalaryDecreesResponsePagedResult"
              }
            ]
          },
          "errors": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "UpdateSalaryDecreeRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "employeeId": {
            "type": "string",
            "format": "uuid"
          },
          "salaryProfile": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/SalaryDecreeRequest"
              }
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter token without Bearer",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
...
}